requirements.txt
Rx~=1.6.3
pip~=24.2
wheel~=0.44.0
zipp~=3.21.0
MarkupSafe~=3.0.2
werkzeug~=3.1.3
click~=8.1.7
jinja2~=3.1.4
blinker~=1.9.0
itsdangerous~=2.2.0
mysql-connector-python~=9.1.0
setuptools~=74.1.2
graphql-core~=2.3.2
six~=1.17.0
promise~=2.3
python-dateutil~=2.9.0.post0
graphene~=2.1.9
graphql-relay~=2.0.1
aniso8601~=7.0.0
flask~=3.1.0
packaging~=24.1
tomli~=2.0.1
autocommand~=2.2.2
inflect~=7.3.1
more-itertools~=10.3.0
typeguard~=4.3.0
importlib-metadata~=8.5.0
platformdirs~=4.2.2
importlib-resources~=6.4.0
Flask-Cors~=5.0.0